# By wangli (9) and others
# Via wangli (7) and others
* 'master' of https://github.com/nicolastinkl/AISBOSS: (22 commits)
  增加wished 和noted
  修改金额
  修改上传参数的构造逻辑
  修复 proposal 详情页面 page 点点消失的bug
  增加wish_id
  强制默认英语
  增加切换用户通知
  修复 录音时间过短的多语言bug
  修改暂存参数类型不正确bug
  fix 报文
  删除不用的方法
  上传参数
  修改更换用户后重新加载数据的逻辑
  fix
  增加接口
  修复 多次切换语言失灵的问题
  添加默认选中参数设置逻辑
  修改时间和默认项
  修改孕期疗养数据提交不正确的bug
  修复上传数据为空的bug
  ...

Conflicts:
	AISBOSS/Models/ProposalModel.h
	AISBOSS/ViewController/Buyer/AIServiceContentViewController.swift

+++ AISBOSS/Models/ProposalModel.h
@@ -7,7 +7,7 @@
 //
 
 #import "JSONModel.h"
-
+#import <UIKit/UIKit.h>
 
 @protocol ServiceOrderModel
 @end
@@ -109,7 +109,7 @@
 @property (strong, nonatomic) NSString<Optional> * discount;
 @property (strong, nonatomic) NSString<Optional> * saved;
 
-@property (assign, nonatomic) float price;
+@property (assign, nonatomic) CGFloat price;
 @property (strong, nonatomic) NSString<Optional> * billing_mode;
 @property (strong, nonatomic) NSString<Optional> * unit;
 @end
